requirements.txt
absl-py==0.8.1
alabaster==0.7.12
alembic==1.3.1
altair==4.1.0
anaconda-client==1.7.2
anaconda-project==0.8.3
asn1crypto==1.2.0
astor==0.8.0
astroid==2.3.3
astropy==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.8.1
bitarray==1.1.0
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bm3d==3.0.7
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.6.0.post1
cryptography==2.8
cvxopt==1.2.4
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.8.1
dataset==1.1.2
decorator==4.4.1
defusedxml==0.6.0
distributed==2.8.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
fsspec==0.6.0
gast==0.2.2
gevent==1.4.0
gitdb==4.0.7
GitPython==3.1.24
glob2==0.7
google-pasta==0.2.0
graphviz==0.16
greenlet==0.4.15
grpcio==1.16.1
h5py==2.8.0
HeapDict==1.0.1
html5lib==1.0.1
idna==2.8
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==0.23
intel-openmp==2019.0
ipdb==0.12.2
ipykernel==5.1.3
ipython==7.9.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
llvmlite==0.30.0
locket==0.2.0
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl==2019.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.2
numba==0.46.0
numexpr==2.7.0
numpy==1.19.3
numpydoc==0.9.1
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.1
opt-einsum==3.3.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
partd==1.1.0
path.py==12.0.2
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.2.1
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.19.0
psutil==5.6.5
py==1.8.0
pyarrow==5.0.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydeck==0.7.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.4
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyparsing==2.4.5
pypng==0.0.20
pyreadline==2.1
pyrsistent==0.15.6
PySocks==1.7.1
pytest==5.3.0
pytest-arraydiff==0.3
pytest-astropy==0.6.0
pytest-astropy-header==0.1.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
ruamel_yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six @ file:///C:/ci/six_1605205426665/work
smmap==4.0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.11
statsmodels==0.10.1
streamlit==1.0.0
sympy==1.4
tables==3.4.4
tb-nightly==1.14.0a20190603
tblib==1.5.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019060501
tifffile==2020.10.1
toml==0.10.2
toolz==0.10.0
torch==1.2.0
torchvision==0.4.1
torchviz==0.0.1
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.3
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
unicodecsv==0.14.1
urllib3==1.24.2
validators==0.18.2
watchdog==2.1.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.6
xlwings==0.16.0
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
